+++ src/app/src/util/util.facilitiesCSV.js
@@ -1,5 +1,6 @@
 /* eslint-disable camelcase */
 import isArray from 'lodash/isArray';
+import get from 'lodash/get';
 
 import { joinDataIntoCSVString } from './util';
 import { PPE_FIELD_NAMES } from './constants';
@@ -24,13 +25,14 @@ export const createFacilityRowFromFeature = (feature, options) => {
             oar_id,
             contributors,
             is_closed,
+            contributor_fields,
         },
         geometry: {
             coordinates: [lng, lat],
         },
     } = feature;
 
-    const contributorFields =
+    const contributorData =
         options && options.isEmbedded
             ? []
             : [contributors ? contributors.map(c => c.name).join('|') : ''];
@@ -48,19 +50,29 @@ export const createFacilityRowFromFeature = (feature, options) => {
         options && options.includeClosureFields
             ? [is_closed ? 'CLOSED' : null]
             : [];
+    const contributorFields =
+        options && options.isEmbedded
+            ? contributor_fields.map(field => field.value)
+            : [];
 
     return Object.freeze(
         [oar_id, name, address, country_code, country_name, lat, lng]
-            .concat(contributorFields)
+            .concat(contributorData)
             .concat(ppeFields)
-            .concat(closureFields),
+            .concat(closureFields)
+            .concat(contributorFields),
     );
 };
 
 export const makeFacilityReducer = options => (acc, next) =>
     acc.concat([createFacilityRowFromFeature(next, options)]);
 
-export const makeHeaderRow = options => {
+const getContributorFieldHeaders = facilities => {
+    const fields = get(facilities, '[0].properties.contributor_fields', []);
+    return fields.map(field => field.label);
+};
+
+export const makeHeaderRow = (options, facilities = []) => {
     let headerRow = csvHeaders;
     if (!options || !options.isEmbedded) {
         headerRow = headerRow.concat(['contributors']);
@@ -71,11 +83,17 @@ export const makeHeaderRow = options => {
     if (options && options.includeClosureFields) {
         headerRow = headerRow.concat(['is_closed']);
     }
+    if (options && options.isEmbedded) {
+        headerRow = headerRow.concat(getContributorFieldHeaders(facilities));
+    }
     return [headerRow];
 };
 
 export const formatDataForCSV = (facilities, options = {}) =>
-    facilities.reduce(makeFacilityReducer(options), makeHeaderRow(options));
+    facilities.reduce(
+        makeFacilityReducer(options),
+        makeHeaderRow(options, facilities),
+    );
 
 export const createFacilitiesCSV = (facilities, options = {}) => {
     const data = formatDataForCSV(facilities, options);